feat: unified distance-proportional movement + waypoint debug beams
Movement System (float tick cost pipeline):
- WaypointLink.baseTickCost: int -> float (distance * 3f, no rounding)
- Pathfinder.totalTicks: int -> float
- SimEvent.tickCost: int -> float
- SimulationManager: removed WalkSpeed, added TicksPerWorldUnit (3f)
- Duration now coupled: duration = linkCost * tickInterval
- Rounding only at budget boundary (ticksElapsed/ticksRemaining)
- Fix exit zip: compute distance-proportional cost for pathIndex==0
- Entry/exit moves use same coupled formula

Animation:
- Set consistent walkSpeed (1.0f) on each StartMove
- autoDetectWalking already disabled in Initialize

Waypoint Debug (F7):
- PathDebugRenderer: showWaypointGraph flag + SetDebugGraph()
- Renders all graph links/nodes as beams in Game view
- F7 toggles beams in both planning and execution modes
- PDR created during planning phase (not just execution)
- Fix path beam consumption: use visualPathIndex instead of simManager.PathIndex

Camera:
- Zoom speed doubled (ortho 0.25->0.5, perspective 0.15->0.3)
- F6/F7 hotkeys moved before IsExecutionMode early return

+++ Assets/Scripts/Sim/SimEventStream.cs
@@ -28,7 +28,7 @@ namespace SteelCity.Sim
         public Vector3 fromPos;
         public Vector3 toPos;
         public float duration;
-        public int tickCost;
+        public float tickCost;
         public string linkType;
         public string nodeId;
         public string blockId;
@@ -41,7 +41,7 @@ namespace SteelCity.Sim
         public int dialogTotalTicks;
         public string message;
 
-        public static SimEvent Move(Vector3 from, Vector3 to, float duration, int tickCost, string linkType,
+        public static SimEvent Move(Vector3 from, Vector3 to, float duration, float tickCost, string linkType,
                                      string nodeId, int tickElapsed, int tickRemaining)
         {
             return new SimEvent
